refactor backend

+++ backend/api/relayer.py
@@ -1,6 +1,7 @@
 from playhouse.shortcuts import model_to_dict
+from peewee import ProgrammingError
 from model import Relayer
-from exception import InvalidValueException
+from exception import InvalidValueException, MissingArgumentException
 from .base import BaseHandler
 
 
@@ -14,22 +15,42 @@ class RelayerHandler(BaseHandler):
     async def post(self):
         """Add new relayer"""
         relayer = self.request_body
-        obj = await self.application.objects.create(Relayer, **relayer)
-        self.json_response(model_to_dict(obj))
+        try:
+            obj = await self.application.objects.create(Relayer, **relayer)
+            self.json_response(model_to_dict(obj))
+        except Exception as err:
+            raise InvalidValueException('relayer payload is invalid: {param}'.format(param=str(relayer)))
 
     async def patch(self):
         """Update existing relayer"""
         relayer = self.request_body
-        relayer_id = relayer.get('id')
+        relayer_id = relayer.get('id', None)
+
+        if not relayer_id:
+            raise MissingArgumentException('missing relayer id')
+
         del relayer['id']
-        query = (Relayer.update(**relayer).where(Relayer.id == relayer_id).returning(Relayer))
-        cursor = query.execute()
-        obj = cursor[0]
-        self.json_response(model_to_dict(obj))
+
+        try:
+            query = (Relayer.update(**relayer).where(Relayer.id == relayer_id).returning(Relayer))
+            cursor = query.execute()
+            obj = cursor[0]
+            self.json_response(model_to_dict(obj))
+        except IndexError as err:
+            raise InvalidValueException('relayer id={param} does not exist'.format(param=str(relayer_id)))
+        except ProgrammingError as err:
+            raise InvalidValueException('update payload is invalid: {param}'.format(param=str(relayer)))
 
     async def delete(self):
         """Delete a relayer"""
-        relayer_id = self.get_argument('id')
-        relayer = Relayer.get(Relayer.id == relayer_id)
-        result = relayer.delete_instance()
-        self.json_response({})
+        relayer_id = self.get_argument('id', None)
+
+        if not relayer_id:
+            raise MissingArgumentException('missing relayer id')
+
+        try:
+            relayer = Relayer.get(Relayer.id == relayer_id)
+            relayer.delete_instance()
+            self.json_response({})
+        except Exception as err:
+            raise InvalidValueException('relayerId: {param} -- {error}'.format(param=relayer_id, error=err.__class__.__name__))
